% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDsetList.R
\name{compareDsetList}
\alias{compareDsetList}
\title{Column wise comparison of a dataset to a list of datasets.}
\usage{
compareDsetList(x, y)
}
\arguments{
\item{x}{the path to a normalized dataset x}

\item{y}{a character vector of all paths to datasets to compare}
}
\value{
a character vector of all datasets for which dataset x shares at least one sample, separated by ;
}
\description{
This function finds all datasets of the list y for which dataset x shares samples. The datasets are in the form
variables (probes) x samples. The number of variables (probes) in both datasets should be the same.
Let us assume we want to compare normalized gene expression dataset GSE86013 with datasets:
GSE86015, GSE9008, GSE9119. x and y can be either local paths where .Rda normalized data are stored
or links to the csv files in BioDataome.
First example runs with datasets stored in .csv in BioDataome.
Since these datasets are large we propose to use fread from package data.table to read datasets faster
}
\examples{
x<-"http://dataome.mensxmachina.org/data/Homo\%20sapiens/GPL570/GSE86013.csv"
y<-c("GSE86015.csv","GSE9008.csv","GSE9119.csv")
y<-paste0("http://dataome.mensxmachina.org/data/Homo\%20sapiens/GPL570/",y)
commonGSEs<-compareDsetList(x,y)
}
