% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDsets.R
\name{compareDsets}
\alias{compareDsets}
\title{Column wise comparison of two datasets
This function finds how many samples are shared between two datasets. The datasets are in the form
variables (probes) x samples. The number of variables (probes) in both datasets should be the same
Let us assume we want to compare two normalized gene expression datasets from the same platform}
\usage{
compareDsets(d1, d2)
}
\arguments{
\item{d1}{a numeric matrix of a dataset}

\item{d2}{a numeric matrix of a dataset}
}
\value{
the number of equal samples
}
\description{
Column wise comparison of two datasets
This function finds how many samples are shared between two datasets. The datasets are in the form
variables (probes) x samples. The number of variables (probes) in both datasets should be the same
Let us assume we want to compare two normalized gene expression datasets from the same platform
}
\examples{
\dontrun{
d1<-get(load(url("http://dataome.mensxmachina.org/data/Homo\%20sapiens/GPL570/GSE86013.Rda")))
d2<-get(load(url("http://dataome.mensxmachina.org/data/Homo\%20sapiens/GPL570/GSE86015.Rda")))
commons<-compareDsets(d1,d2) }

}
