% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlSamples.R
\name{controlSamples}
\alias{controlSamples}
\title{Discover control samples from phenotype data in GEO}
\usage{
controlSamples(d)
}
\arguments{
\item{d}{a data frame with the contents of series matrix}
}
\value{
a data frame of GEO sample ids (i.e. GSM60555) and their class.
}
\description{
This function discovers control samples from the series matrix found in GEO.
It searches for specific keywords that are often used to denote controls in specific columns of the
series matrices.
}
\examples{
\dontrun{
phenos<-downloadPhenotypePlatform("GSE11761","GPL570")}
controls<-controlSamples(BioDataome:::phenos)
}
