% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curateGSE.R
\name{curateGSE}
\alias{curateGSE}
\title{Run all steps to download, preprocess and annotate a GEO dataset}
\usage{
curateGSE(x, y, z)
}
\arguments{
\item{x}{a GSE series ID}

\item{y}{a GEO platform id (GPL)}

\item{z}{the path to write the output}
}
\value{
writes in the given path two data frames, the preprocessed data and the metadata file with phenotype information
}
\description{
Given a GSE id this function downloads, preprocesses, annotates a study and also creates the metadata. It saves two files,
the data file and the GEO metadata file in the given path.
}
\examples{
curateGSE("GSE11761","GPL570",getwd())
}
