% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curateRecountRNASeq.R
\name{curateRecountRNASeq}
\alias{curateRecountRNASeq}
\title{Run all steps to download, preprocess and annotate an RNASeq dataset from Recount}
\usage{
curateRecountRNASeq(x, y)
}
\arguments{
\item{x}{a recount dataset ID}

\item{y}{the path to write the output}
}
\value{
writes in the given path two data frames, the preprocessed data and the metadata file with phenotype information
}
\description{
Run all steps to download, preprocess and annotate an RNASeq dataset from Recount
}
\examples{
curatedRecount<-curateRecountRNASeq("SRP032775",getwd())
}
