% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasetoChildrenNodes.R
\name{diseasetoChildrenNodes}
\alias{diseasetoChildrenNodes}
\title{Map a Disease Ontology (D-O) term to the first children nodes in D-O}
\usage{
diseasetoChildrenNodes(x)
}
\arguments{
\item{x}{a disease in D-O terms}
}
\value{
the first children node of x disease
}
\description{
This function uses internal look up data to map a disease to its first children node.
}
\examples{
DOChild<-diseasetoChildrenNodes("vesiculitis")
}
