% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasetoParentNodes.R
\name{diseasetoParentNodes}
\alias{diseasetoParentNodes}
\title{Map a Disease Ontology (D-O) term to the parent nodes in D-O}
\usage{
diseasetoParentNodes(x)
}
\arguments{
\item{x}{a disease in D-O terms}
}
\value{
the parent node of x disease
}
\description{
This function uses internal look up data to map a disease to its parent node.
}
\examples{
DOParent<-diseasetoParentNodes("vesiculitis")
}
