% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPhenotypePlatform.R
\name{downloadPhenotypePlatform}
\alias{downloadPhenotypePlatform}
\title{Download series matrix from GEO for a given study and platform}
\usage{
downloadPhenotypePlatform(x, y)
}
\arguments{
\item{x}{a GEO Series id (GSE)}

\item{y}{a GEO platform id (GPL)}
}
\value{
a data frame with the contents of the series matrix found in GEO
}
\description{
This function downloads returns the series matrix related to a given GEO Series (GSE)
}
\examples{
downloadPhenotypePlatform("GSE11761","GPL570")
}
