% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRecount.R
\name{downloadRecount}
\alias{downloadRecount}
\title{Download gene-level RangedSummarizedExperiment data from Recount}
\usage{
downloadRecount(x)
}
\arguments{
\item{x}{a recount dataset ID}
}
\value{
RangedSummarizedExperiment object for the given study
}
\description{
This function downloads the RangedSummarizedExperiment object with the data summarized at the gene level
from Recount (https://jhubiostatistics.shinyapps.io/recount/)
}
\examples{
\dontrun{
downloadRecount("SRP032775",getwd()) }
}
