% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEtoDisease.R
\name{GSEtoDisease}
\alias{GSEtoDisease}
\title{Annotate a study (GSE) with a disease term from the Disease Ontology by exploiting both PubTator and GEO}
\usage{
GSEtoDisease(GSE)
}
\arguments{
\item{GSE}{a GSE series ID}
}
\value{
a character vector of all related diseases, separated by ;
}
\description{
Given a GSE id this function annotates the study with a disease term from the Disease Ontology (D-O):
http://disease-ontology.org/
It provides the most specific disease term, meaning the term with the highest depth in the D-O.
}
\examples{
\dontrun{
diseases<-GSEtoDisease("GSE10245") }
}
