% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessGEOMethylation.R
\name{preprocessGEOMethylation}
\alias{preprocessGEOMethylation}
\title{Preprocess IDAT files from Illumina HumanMethylation450 BeadChip}
\usage{
preprocessGEOMethylation(x)
}
\arguments{
\item{x}{a character array with the paths to the idat files}
}
\value{
a matrix of dimensions: probes x samples with the normalized methylation values
}
\description{
This function utilizes minfi Package to convert data into methylation measurements.
}
\examples{
Assuming there is a directory named GSE78279 in the working directory where idat files are stored and
"GSM2071074_8655685078_R03C02" and "GSM2071074_8655685078_R03C02" are the file names for the idat files,
then x should be:
x<-c("GSM2071074_8655685078_R03C02","GSM2071075_8655685078_R04C02")
x<-file.path(getwd(),"GSE78279",x)
dataNorm<-preprocessGEOMethylation(x)
}
