% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recountIDtoGSE.R
\name{recountIDtoGSE}
\alias{recountIDtoGSE}
\title{Map a recount dataset ID to GSE ID}
\usage{
recountIDtoGSE(x)
}
\arguments{
\item{x}{a recount dataset ID}
}
\value{
a GSE series ID
}
\description{
Map a recount dataset ID to GSE ID
}
\examples{
\dontrun{
recountIDtoGSE("SRP032775") }
}
