% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curateGSE.R
\name{curateGSE}
\alias{curateGSE}
\title{Run all steps to download, preprocess and annotate a GEO dataset}
\usage{
curateGSE(x, y, z = getwd(), keepRaw = FALSE)
}
\arguments{
\item{x}{a GSE series ID}

\item{y}{a GEO technology id (GPL)}

\item{z}{the path to save the downloaded files. By default this value is set to the working directory.}
}
\value{
a list with three components:
\itemize{
 \item{"metadata"}{a data frame of the metadata with sample phenotype information}
 \item{"dataNorm"}{a numeric matrix of the preprocessed data with variables (probes) as rows and samples as columns}
}
}
\description{
Given a GSE id this function downloads, preprocesses, annotates a study and also creates the sample phenotype metadata.
}
\examples{
curateGSE("GSE11761","GPL570",getwd())
}
