% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRaw.R
\name{downloadRaw}
\alias{downloadRaw}
\title{Download raw CEL files from GEO for a given study}
\usage{
downloadRaw(x, y)
}
\arguments{
\item{x}{a GEO Series id (GSE)}

\item{y}{the path to save the downloaded files}
}
\value{
a directory with the GSE with the compressed RAW files
}
\description{
Download raw CEL files from GEO for a given study
}
\examples{
downloadRaw("GSE11761")
}
