% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRaw.R
\name{downloadRaw}
\alias{downloadRaw}
\title{Download raw files from GEO for a given study}
\usage{
downloadRaw(x, y = getwd())
}
\arguments{
\item{x}{a GEO Series id (GSE)}

\item{y}{the path to save the downloaded files. By default this value is set to the working directory}
}
\value{
downloadRaw creates a directory in the given path with the GSE name and saves there the compressed RAW files.
}
\description{
Download raw files from GEO for a given study
}
\examples{

Download the raw files of GSE11761 study and save them in a directory named GSE11761 in the working directory

downloadRaw("GSE11761",getwd())

}
