% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezIDtoGSE.R
\name{entrezIDtoGSE}
\alias{entrezIDtoGSE}
\title{Find all GSE ids for a given Entrez query
Example: query GEO for all Homo sapiens studies with sample size
between 200-300, measured with GPL570 and provide CEL files}
\usage{
entrezIDtoGSE(x)
}
\arguments{
\item{x}{an esearch object as a result of an entrez_search query}
}
\value{
a matrix the first column of which is the GSE id and the second the entrezID
}
\description{
Find all GSE ids for a given Entrez query
Example: query GEO for all Homo sapiens studies with sample size
between 200-300, measured with GPL570 and provide CEL files
}
\examples{
r_search <- rentrez::entrez_search(db="gds", term="Homo sapiens[ORGN]
AND CEL[SFIL] AND gpl570[ACCN] AND 200:300[Number of Samples]",
retmax =10000, use_history=TRUE)
\dontrun{
entrezIDtoGSE(r_search) }

}
