% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEmetadata.R
\name{GSEmetadata}
\alias{GSEmetadata}
\title{Create metadata of a GEO dataset for BioDataome}
\usage{
GSEmetadata(x, y)
}
\arguments{
\item{x}{a GSE series ID}

\item{y}{a GEO platform id (GPL)}
}
\value{
a data frame of metadata with columns: sample IDs, Class and all other GEO phenotype data found in series matrix
}
\description{
Given a GSE id this function downloads phenotype data from GEO for a specific study, discovers control samples
and creates the metadata file for BioDataome
}
\examples{
metadata<-GSEmetadata("GSE11761","GPL570")
}
